pyproject.toml
[project]
name = "model-informed-greenhouse-dashboard"
version = "0.1.0"
description = "Model-informed greenhouse dashboard"
authors = [
    {name = "Phytoritas",email = "phmoon08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.109,<0.110",
    "uvicorn[standard]>=0.27,<0.28",
    "pydantic>=2.11,<3.0",
    "pydantic-settings>=2.10,<3.0",
    "python-dotenv>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
    "pandas>=2.1,<3.0",
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<2.0",
    "watchdog>=3.0,<4.0",
    "httpx>=0.27,<0.28",
    "openai (>=2.30.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
ruff = "^0.15.8"

